third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-comparison-dom-groups-change-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that Comparison view of heap snapshots will contain added nodes even if their ids are less than the maximumm JS object id in the base snapshot.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
Delta: +4 -1
Deleted node id(s): 40
Added node id(s): 15,25,35,5

Profiler was disabled.
